Translator 内で GAS 製の翻訳機能に依存する部分を別クラスに切り出し
